Resolve Login Regis

+++ routes/web.php
@@ -27,7 +27,6 @@ Route::get('/kegiatan', [PageController::class, 'kegiatan'])->name('kegiatan');
 Route::get('/organisasi', [PageController::class, 'organisasi'])->name('organisasi');
 
 // Auth Routes
-// Auth Routes
 Route::get('/login', [AuthController::class, 'showLogin'])->name('login');
 Route::post('/login', [AuthController::class, 'login']);
 Route::get('/register', [AuthController::class, 'showRegister'])->name('register');
@@ -36,21 +35,19 @@ Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
 Route::get('admin/mass-registration', [MassRegistrationController::class, 'showForm'])->name('admin.mass.register.form');
 Route::post('admin/mass-registration', [MassRegistrationController::class, 'processForm'])->name('admin.mass.register');
 
-// Ganti route dashboard yang lama dengan yang baru
 Route::middleware(['auth'])->group(function () {
     Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
-    Route::resource('events', EventController::class);
+    Route::resource('events', EventController::class)->except(['create', 'store', 'edit', 'update', 'destroy']);
     Route::get('/events/{event}/scan', [EventController::class, 'showScanner'])->name('events.scan');
     Route::post('/events/attendance', [EventController::class, 'processAttendance'])->name('events.attendance');
 });
 
-// Tambahkan middleware untuk admin
-Route::middleware(['auth', 'admin'])->group(function () {
-    Route::get('events/create', [EventController::class, 'create'])->name('events.create');
-    Route::post('events', [EventController::class, 'store'])->name('events.store');
-    Route::get('events/{event}/edit', [EventController::class, 'edit'])->name('events.edit');
-    Route::put('events/{event}', [EventController::class, 'update'])->name('events.update');
-    Route::delete('events/{event}', [EventController::class, 'destroy'])->name('events.destroy');
-    Route::resource('bidang', BidangController::class);
-    Route::get('/admin/users', [UserController::class, 'index'])->name('users.index');
+Route::middleware(['auth', 'admin'])->prefix('admin')->group(function () {
+    Route::get('events/create', [EventController::class, 'create'])->name('admin.events.create');
+    Route::post('events', [EventController::class, 'store'])->name('admin.events.store');
+    Route::get('events/{event}/edit', [EventController::class, 'edit'])->name('admin.events.edit');
+    Route::put('events/{event}', [EventController::class, 'update'])->name('admin.events.update');
+    Route::delete('events/{event}', [EventController::class, 'destroy'])->name('admin.events.destroy');
+
+    Route::get('/users', [UserController::class, 'index'])->name('users.index');
 });
